fix(review): keep maintenance argv contract below smoke utilities
